dev/requirements.txt
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
easyocr==1.7.1
executing==2.1.0
filelock==3.16.0
fsspec==2024.9.0
idna==3.9
imageio==2.35.1
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numpy==2.1.1
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.3
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pyclipper==1.3.0.post5
Pygments==2.18.0
pytesseract==0.3.13
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
six==1.16.0
stack-data==0.6.3
sympy==1.13.2
tifffile==2024.8.30
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
